% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-benchmark.R
\name{bench_combine}
\alias{bench_combine}
\title{Combine two benchmarks into one}
\usage{
bench_combine(current, new)
}
\arguments{
\item{current}{current benchmark}

\item{new}{new benchmark}
}
\value{
benchmark object with names "current" and "new" appended to expression.
}
\description{
Combine two benchmarks into one
}
\examples{
dat <- data.frame(x = runif(100, 1, 1000), y = runif(10, 1, 1000))
bm1 <- bench::mark(
    subset = dat[dat$x > 500, ]
)

bm2 <- bench::mark(
    subset = dat[which(dat$x > 500), ]
)

combined_bm <- bench_combine(
    bm1,
    bm2
)
}
