% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-benchmark.R
\name{bench_compare}
\alias{bench_compare}
\title{Compare benchmark results}
\usage{
bench_compare(benchmark)
}
\arguments{
\item{benchmark}{A benchmark.}
}
\value{
A combined benchmark object.
}
\description{
Compare benchmark results
}
