% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{bench_path}
\alias{bench_path}
\title{Get path to benchmark file}
\usage{
bench_path(benchmark)
}
\arguments{
\item{benchmark}{A benchmark.}
}
\value{
A file path string to the benchmark RDS file.
}
\description{
Get path to benchmark file
}
