% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Run benchmarks and save results}
\usage{
benchmark(name = NULL)
}
\arguments{
\item{name}{Optional function name. If NULL, runs all benchmarks.}
}
\description{
Run benchmarks and save results
}
