% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{branchmark_baseline}
\alias{branchmark_baseline}
\title{Create or update performance baselines}
\usage{
branchmark_baseline(name = NULL, force = FALSE)
}
\arguments{
\item{name}{Optional function name. If NULL, processes all benchmarks.}

\item{force}{Whether to overwrite existing baselines without confirmation}
}
\description{
Create or update performance baselines
}
