% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{branchmark_compare}
\alias{branchmark_compare}
\title{Compare current benchmarks with baselines}
\usage{
branchmark_compare(name = NULL, threshold = 0.1)
}
\arguments{
\item{name}{Optional function name to compare. If NULL, compares all.}

\item{threshold}{Performance degradation threshold (default 0.1 = 10\%)}
}
\description{
Compare current benchmarks with baselines
}
