% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{branchmark_snapshot}
\alias{branchmark_snapshot}
\title{Create a snapshot of current benchmark results}
\usage{
branchmark_snapshot(name = NULL)
}
\arguments{
\item{name}{Function name to snapshot}
}
\description{
Create a snapshot of current benchmark results
}
