% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_performance_change}
\alias{calculate_performance_change}
\title{Calculate performance change between current and baseline}
\usage{
calculate_performance_change(current, baseline)
}
\description{
Calculate performance change between current and baseline
}
\keyword{internal}
