% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{check_environment}
\alias{check_environment}
\title{Check current environment and compare with baseline}
\usage{
check_environment(warn = TRUE)
}
\arguments{
\item{warn}{Whether to issue warnings for significant changes}
}
\description{
Check current environment and compare with baseline
}
