% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_all_benchmarks}
\alias{compare_all_benchmarks}
\title{Compare all benchmarks with their baselines}
\usage{
compare_all_benchmarks(threshold)
}
\description{
Compare all benchmarks with their baselines
}
\keyword{internal}
