% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_all_branches}
\alias{compare_all_branches}
\title{Compare all benchmarks between branches}
\usage{
compare_all_branches()
}
\description{
Compare all benchmarks between branches
}
\keyword{internal}
