% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{compare_environments}
\alias{compare_environments}
\title{Compare two environment snapshots}
\usage{
compare_environments(current, baseline)
}
\description{
Compare two environment snapshots
}
\keyword{internal}
