% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_single_benchmark}
\alias{compare_single_benchmark}
\title{Compare a single benchmark with its baseline}
\usage{
compare_single_benchmark(name, threshold)
}
\description{
Compare a single benchmark with its baseline
}
\keyword{internal}
