% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_single_branch}
\alias{compare_single_branch}
\title{Compare a single benchmark between branches}
\usage{
compare_single_branch(name)
}
\description{
Compare a single benchmark between branches
}
\keyword{internal}
