% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{copy_workflow_template}
\alias{copy_workflow_template}
\title{Copy workflow template to destination}
\usage{
copy_workflow_template(template_name, dest_dir, overwrite)
}
\description{
Copy workflow template to destination
}
\keyword{internal}
