% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_example_benchmark}
\alias{create_example_benchmark}
\title{Create example benchmark file}
\usage{
create_example_benchmark(example_path)
}
\description{
Create example benchmark file
}
\keyword{internal}
