% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{load_baseline_environment}
\alias{load_baseline_environment}
\title{Load baseline environment if available}
\usage{
load_baseline_environment()
}
\description{
Load baseline environment if available
}
\keyword{internal}
