% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_bench_rds}
\alias{read_bench_rds}
\title{Read benchmark RDS file}
\usage{
read_bench_rds(benchmark)
}
\arguments{
\item{benchmark}{A benchmark.}
}
\value{
The benchmark object loaded from the corresponding RDS file.
}
\description{
Read benchmark RDS file
}
