% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmiss.R
\name{rowmiss}
\alias{rowmiss}
\title{Calculate ross missing}
\usage{
rowmiss(data)
}
\arguments{
\item{data}{data frame}
}
\value{
number of missing values
}
\description{
Calculate ross missing
}
\examples{
rowmiss(airquality)
}
