% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{run_current_benchmark}
\alias{run_current_benchmark}
\title{Run current benchmark without saving}
\usage{
run_current_benchmark(name)
}
\description{
Run current benchmark without saving
}
\keyword{internal}
