% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{setup_workflows}
\alias{setup_workflows}
\title{Set up GitHub workflows with custom configuration}
\usage{
setup_workflows(runner_os = "ubuntu-latest", r_version = "release")
}
\arguments{
\item{runner_os}{Operating system for GitHub runners ("ubuntu-latest", "macos-latest", "windows-latest")}

\item{r_version}{R version to use in workflows}
}
\description{
Set up GitHub workflows with custom configuration
}
