% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{update_baseline}
\alias{update_baseline}
\title{Update a specific baseline}
\usage{
update_baseline(name, force = FALSE)
}
\arguments{
\item{name}{Function name}

\item{force}{Whether to overwrite without confirmation}
}
\description{
Update a specific baseline
}
