% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_benchmark}
\alias{use_benchmark}
\title{Create a benchmark template for a function}
\usage{
use_benchmark(name, open = rlang::is_interactive())
}
\arguments{
\item{name}{Function name to create benchmark for}

\item{open}{Whether to open the file for editing}
}
\description{
Create a benchmark template for a function
}
