% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-benchthat.R
\name{use_benchthat}
\alias{use_benchthat}
\title{Sets up overall benchmarking infrastructure}
\usage{
use_benchthat()
}
\description{
Creates \verb{benchmarks/benchthat/}, \code{benchmarks/benchthat.R}, and adds
the benchthat package to the Suggests field. Structure of code borrowed from
\code{\link[usethis:use_r]{usethis::use_test()}}.
}
\examples{
\dontrun{
use_benchthat()

use_bench()

use_bench("function-name")
}
}
\seealso{
\code{\link[=use_bench]{use_bench()}} to create individual benchmark files
}
