% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{use_branchmark_github}
\alias{use_branchmark_github}
\title{Set up GitHub Actions workflows for branchmark}
\usage{
use_branchmark_github(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite existing workflow files}
}
\description{
Set up GitHub Actions workflows for branchmark
}
