% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_branchmark_gitignore}
\alias{use_branchmark_gitignore}
\title{Update .gitignore for branchmark}
\usage{
use_branchmark_gitignore(path)
}
\description{
Update .gitignore for branchmark
}
\keyword{internal}
