% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_single_function_benchmark}
\alias{validate_single_function_benchmark}
\title{Validate that benchmark tests only one function}
\usage{
validate_single_function_benchmark(code)
}
\description{
Validate that benchmark tests only one function
}
\keyword{internal}
