% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{warn_system_changes}
\alias{warn_system_changes}
\title{Warn about system changes that might affect benchmarks}
\usage{
warn_system_changes(changes)
}
\arguments{
\item{changes}{List of environment changes}
}
\description{
Warn about system changes that might affect benchmarks
}
