% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{write_bench_rds}
\alias{write_bench_rds}
\title{Write benchmark to RDS}
\usage{
write_bench_rds(benchmark)
}
\arguments{
\item{benchmark}{A benchmark object.}
}
\value{
No return value, called for side effects.
}
\description{
Write benchmark to RDS
}
