% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{branchmark}
\alias{branchmark}
\title{Run benchmarks and compare with other branch}
\usage{
branchmark(name = NULL)
}
\arguments{
\item{name}{Optional function name. If NULL, compares all benchmarks.}
}
\description{
Run benchmarks and compare with other branch
}
