% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_benchmark}
\alias{use_benchmark}
\title{Create a benchmark test for a function}
\usage{
use_benchmark(name, open = rlang::is_interactive())
}
\arguments{
\item{name}{Name of the function to benchmark. Can be quoted or unquoted.}

\item{open}{Whether to open the created file for editing}
}
\description{
Creates a benchmark file for a specific function, similar to usethis::use_test().
The benchmark file will be created in the branchmark/ directory with appropriate
template code for testing the specified function.
}
\examples{
\dontrun{
use_benchmark("my_function")
use_benchmark(another_function)
}
}
